Fixed ds_x and ds_y offsets in all back-ends.

+++ src/openslide-ops-ngr.c
@@ -170,8 +170,8 @@ static void paint_region(openslide_t *osr, cairo_t *cr,
 
   // compute coordinates
   double ds = openslide_get_layer_downsample(osr, layer);
-  int64_t ds_x = x / ds;
-  int64_t ds_y = y / ds;
+  double ds_x = x / ds;
+  double ds_y = y / ds;
   int64_t start_tile_x = ds_x / ngr->column_width;
   int64_t end_tile_x = ((ds_x + w) / ngr->column_width) + 1;
   int64_t start_tile_y = ds_y;
