+++ src/openslide-ops-tiff.c
@@ -358,8 +358,8 @@ static void paint_region_unlocked(openslide_t *osr, cairo_t *cr,
   }
 
   double ds = openslide_get_layer_downsample(osr, layer);
-  int64_t ds_x = x / ds;
-  int64_t ds_y = y / ds;
+  double ds_x = x / ds;
+  double ds_y = y / ds;
   int64_t start_tile_x = ds_x / (tw - ox);
   int64_t end_tile_x = ((ds_x + w) / (tw - ox)) + 1;
   int64_t start_tile_y = ds_y / (th - oy);
